+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 Historical list of changes in releases
+## [v0.0.37] - 2024-06-13
+* Add support for custom 'tileActions' ([#64](https://github.com/digitalsamba/embedded-sdk/pull/64)).
+* Add `chatMessageReceived`  event to type definitions ([#ed7cd15](https://github.com/digitalsamba/embedded-sdk/commit/ed7cd15c8e5fdde32935c814e2573f6a4d72ce9d))
+* Add `requireConfirmation` prop to `endSession` command (true by default) ([#e5476c6](https://github.com/digitalsamba/embedded-sdk/commit/e5476c6af06a05d0f4ff8c404e1270073cf65f91))
+
 ## [v0.0.35] - 2023-11-16
 * Add `sessionEnded`  event to type definitions ([#146f159](https://github.com/digitalsamba/embedded-sdk/commit/146f1599e74590aa0a3f130bf38cee1071120874))